requirements.txt
streamlit>=1.18
pandas>=1.3
numpy>=1.21
plotly>=5.0
transformers>=4.30
tensorflow>=2.8
emoji>=2.0.0
datasets>=2.0
requests>=2.25
python-dotenv>=0.21
